refactor(tui): extract the IO event loop into tui/run_loop.rs
Move run_loop, BgTaskOutcome, the gist/local spawn helpers, and the
background IO helpers (pin sync, browser, edit, download, refresh) into a
child run_loop module; move the remaining pure render/diff helpers
(render_centered_modal, time/label formatters, diff_labels, preview_diff_text)
into render.rs. run_loop.rs reaches AppState and the render helpers via
use super::*. run() calls run_loop via a pub(super) export. Pure refactor,
no behaviour change. Part of #73.
